fix validation skipping for null on scalars (#342)
This PR fixes an inconsistent behavior when using `null` in a variable
for a required type (e.g. `String!`)
